backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
crewai>=0.120.0
langchain-google-genai>=2.1.0
duckduckgo-search>=6.2.0
pydantic>=2.8.0
sqlmodel>=0.0.22
SQLAlchemy>=2.0.36
alembic>=1.14.0
psycopg[binary]>=3.2.3
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
cryptography>=43.0.3
email-validator>=2.2.0
python-dotenv>=1.0.1
tenacity>=8.2.3
orjson>=3.10.0
pytest>=8.2.0
pypdf>=4.3.1
python-docx>=1.1.2
python-multipart>=0.0.9
